+++ _example/example02/go.mod
@@ -3,7 +3,7 @@ module example
 go 1.16
 
 require (
-	github.com/golang-queue/queue v0.0.7-0.20210820051048-3cc52a575c1c
+	github.com/golang-queue/queue v0.0.7-0.20210820234255-278ee8112584 // indirect
 	github.com/golang/protobuf v1.5.2 // indirect
 	github.com/nats-io/nats-server/v2 v2.3.2 // indirect
 	github.com/nsqio/go-nsq v1.0.8 // indirect
